fixing more bugs
